+++ src/resources/index.ts
@@ -10,10 +10,16 @@ export {
   type ParseJobListParams,
 } from './parse-jobs';
 export {
+  type ClassifyResponse,
   type ExtractResponse,
+  type ExtractBuildSchemaResponse,
   type ParseResponse,
+  type SectionResponse,
   type SplitResponse,
+  type ClassifyParams,
   type ExtractParams,
+  type ExtractBuildSchemaParams,
   type ParseParams,
+  type SectionParams,
   type SplitParams,
 } from './top-level';